+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## [Unrelease]
 ### Added
-- Add the python code `generatePkgConfig.py` to parse argument from command line
+- Add the python code `generatePkgConfig.py` to use the arguments from command line to find the include path and library files for libtorch.
 
 ### Fixed